requirements.txt
absl-py==1.4.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyconfig==0.10.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astunparse==1.6.3
atomicwrites==1.4.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.0
binaryornot==0.4.4
black==22.12.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
build==0.10.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
confection==0.0.4
cookiecutter==2.1.1
coverage==7.2.2
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dynaconf==3.1.12
entrypoints==0.4
fastjsonschema==2.16.3
flake8==4.0.1
flatbuffers==23.5.9
fsspec==2023.1.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.55.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isort==5.12.0
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.2.0
json5==0.9.11
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.23.6
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.15.2
kedro==0.18.6
kedro-telemetry==0.2.3
keras==2.12.0
langcodes==3.3.0
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.1.0
more-itertools==9.1.0
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nbstripout==0.6.1
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.5.0
pip-tools==6.12.3
platformdirs==3.1.1
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
py==1.11.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.6
pydeck==0.8.1b0
pyflakes==2.4.0
Pygments==2.14.0
Pympler==1.0.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==1.13.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytoolconfig==1.2.5
pytz==2022.7.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.2
rope==1.7.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
streamlit==1.22.0
streamlit-lottie==0.0.3
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
toposort==1.10
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
validators==0.20.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
widgetsnbextension==4.0.5
wrapt==1.14.1
zipp==3.15.0
